create routing model base classes
